core now free from inputs from applications
